% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.me-package.R
\docType{package}
\name{praise.me-package}
\alias{praise.me}
\alias{praise.me-package}
\title{praise.me: You are awesome!}
\description{
The package generates messages for affirmation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/praise.me}
  \item \url{http://emitanaka.org/praise.me/}
  \item Report bugs at \url{https://github.com/emitanaka/praise.me/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

}
\keyword{internal}
