% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise_me}
\alias{praise_me}
\alias{praise_someone}
\title{Praises you or someone}
\usage{
praise_me(by = NULL)

praise_someone(who = NULL, by = NULL)
}
\arguments{
\item{by}{A character to say the praise. See the full
list of character by \code{list_character()}.}

\item{who}{A character of who to praise.}
}
\value{
An object of class \code{cheer}, which is
just a character with special print method.
}
\description{
Praises you or someone with a random word.
}
\examples{
praise_me()
praise_me(by = "cow")
praise_someone()
praise_someone("Joanna", by = "cat")

}
