% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{praises}
\alias{praises}
\title{A list of praises}
\format{
A data frame with a single column.
\describe{
\item{words}{A list of praises.}
}
}
\source{
\url{https://www.vocabulary.com/lists/5167}
}
\usage{
praises
}
\description{
A list of praises
}
\keyword{datasets}
